Add prop-types validation to LockInButton; Reorganize files placing everything that connects to redux store inside containers and everything that doesn't connect inside components
